# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nutrition E-commerce API Documentation",
        "description": "Apis to perform various actions on resources",
        "contact": {
            "email": "support@email.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Test API Server"
        },
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Fetches a paginated list of all available products",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products Fetched Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Products Fetched Successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        },
                                        "pagination": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/PaginationResource"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Creates a new product and returns the created product details",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product Created Successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product",
                "description": "Fetches details of a specific product by ID.",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Student Fetched Successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No Student found"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "Updates an existing product's information",
                "operationId": "c474d1318e5b8c3de1f4d52b6053111f",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of product to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product Updated Successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateProductRequest": {
                "required": [
                    "name",
                    "category_id",
                    "status",
                    "quantity",
                    "price",
                    "weight",
                    "image_url"
                ],
                "properties": {
                    "name": {
                        "description": "Unique product name",
                        "type": "string",
                        "example": "Protein Shake"
                    },
                    "category_id": {
                        "description": "Must exist in the categories table",
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "description": "0 for inactive, 1 for active",
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": 1
                    },
                    "quantity": {
                        "description": "Stock quantity",
                        "type": "string",
                        "example": 50
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "format": "float",
                        "example": 299.99
                    },
                    "weight": {
                        "description": "Product weight",
                        "type": "string",
                        "example": "500g"
                    },
                    "image_url": {
                        "description": "Product image",
                        "type": "string",
                        "example": "https://image.jpeg"
                    }
                },
                "type": "object"
            },
...
            "ProductResource": {
                "title": "ProductResource",
                "description": "Product resource representation",
                "properties": {
                    "id": {
                        "description": "Product ID",
                        "type": "integer",
                        "example": 1
                    },
                    "product_name": {
                        "description": "Product Name",
                        "type": "string",
                        "example": "Whey Protein"
                    },
                    "price": {
                        "description": "Product Price (MRP)",
                        "type": "number",
                        "format": "float",
                        "example": 2999.99
                    },
                    "weight": {
                        "description": "Product Weight",
                        "type": "string",
                        "example": "2kg"
                    },
                    "quantity": {
                        "description": "Available Stock Quantity",
                        "type": "integer",
                        "example": 50
                    },
                    "svp_points": {
                        "description": "SVP Points Earned",
                        "type": "integer",
                        "example": 15
                    },
                    "status": {
                        "description": "Product Status",
                        "type": "string",
                        "enum": [
                            "active",
                            "in-active"
                        ],
                        "example": "active"
                    },
                    "addedOn": {
                        "description": "Product Added Date",
                        "type": "string",
                        "format": "date",
                        "example": "26-02-2025"
                    },
                    "category": {
                        "description": "Product category details",
                        "properties": {
                            "id": {
                                "description": "Category ID",
                                "type": "integer",
                                "example": 5
                            },
                            "name": {
                                "description": "Category Name",
                                "type": "string",
                                "example": "Supplements"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductResource"
                }
            },
            "PaginationResource": {
                "title": "PaginationResource",
                "description": "Pagination meta info resource",
                "properties": {
                    "total": {
                        "type": "integer",
                        "example": 100
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": "http://example.com/api/products?page=2",
                        "nullable": true
                    },
                    "prev_page_url": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "url": {
                        "properties": {
                            "path": {
                                "type": "string",
                                "example": "http://example.com/api/products"
                            },
                            "pageName": {
                                "type": "string",
                                "example": "page"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "apiKey",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}